*extracted eth http related code 'http-primitives'
