+++ Cargo.toml
@@ -18,6 +18,7 @@ members = [
 	"precompiles/utils",
 	"precompiles/utils/macro",
 	"primitives/dao",
+	"primitives/http",
     "runtime",
 ]
