+++ pallets/dao/src/lib.rs
@@ -16,7 +16,6 @@ use serde::{self, Deserialize};
 use sp_core::crypto::KeyTypeId;
 use sp_io::offchain_index;
 use sp_runtime::{
-	offchain,
 	traits::{AccountIdConversion, AtLeast32BitUnsigned, BlockNumberProvider, StaticLookup},
 	transaction_validity::{
 		InvalidTransaction, TransactionSource, TransactionValidity, ValidTransaction,
@@ -25,6 +24,7 @@ use sp_runtime::{
 use sp_std::{prelude::*, str};
 
 use dao_primitives::*;
+use http_primitives::EthRpcProvider;
 
 #[cfg(test)]
 mod mock;
@@ -73,14 +73,6 @@ pub type BlockNumber = u32;
 pub const KEY_TYPE: KeyTypeId = KeyTypeId(*b"sctl");
 
 const UNSIGNED_TXS_PRIORITY: u64 = 100;
-
-const ERC20_TOKEN_TOTAL_SUPPLY_SIGNATURE: &str =
-	"0x18160ddd0000000000000000000000000000000000000000000000000000000000000000";
-const ERC20_TOKEN_BALANCE_OF_SIGNATURE_PREFIX: &str = "0x70a08231000000000000000000000000";
-
-const FETCH_TIMEOUT_PERIOD: u64 = 6000; // in milli-seconds
-const LOCK_TIMEOUT_EXPIRATION: u64 = FETCH_TIMEOUT_PERIOD + 1000; // in milli-seconds
-const LOCK_BLOCK_EXPIRATION: u32 = 3; // in block number
 
 const ONCHAIN_TX_KEY: &[u8] = b"societal-dao::storage::tx";
 
@@ -155,12 +147,10 @@ pub mod pallet {
 		offchain::{AppCrypto, CreateSignedTransaction, SubmitTransaction},
 		pallet_prelude::*,
 	};
-	use serde_json::{json, Value};
+	use http_primitives::EthHttpService;
+	use serde_json::json;
 	use sp_runtime::{
-		offchain::{
-			storage::StorageValueRef,
-			storage_lock::{BlockAndTime, StorageLock},
-		},
+		offchain::storage::StorageValueRef,
 		traits::{Hash, Zero},
 	};
 
@@ -244,6 +234,8 @@ pub mod pallet {
 
 		/// The identifier type for an offchain worker.
 		type AuthorityId: AppCrypto<Self::Public, Self::Signature>;
+
+		type OffchainHttpService: EthHttpService;
 	}
 
 	/// Origin for the dao pallet.
@@ -314,10 +306,9 @@ pub mod pallet {
 				let mut call = None;
 				match offchain_data.clone() {
 					OffchainData::ApproveDao { dao_hash, token_address } =>
-						match Self::parse_token_balance(Self::fetch_token_total_supply(
-							token_address,
-							None,
-						)) {
+						match T::OffchainHttpService::parse_token_balance(
+							T::OffchainHttpService::fetch_token_total_supply(token_address, None),
+						) {
 							Ok(total_supply) => {
 								call =
 									Some(Call::approve_dao { dao_hash, approve: total_supply > 0 });
@@ -335,11 +326,13 @@ pub mod pallet {
 						hash,
 						length_bound: _,
 					} => {
-						let block_number = match Self::parse_block_number(Self::fetch_from_eth(
-							token_address.clone(),
-							Some(json!("eth_blockNumber")),
-							None,
-						)) {
+						let block_number = match T::OffchainHttpService::parse_block_number(
+							T::OffchainHttpService::fetch_from_eth(
+								token_address.clone(),
+								Some(json!("eth_blockNumber")),
+								None,
+							),
+						) {
 							Ok(block_number) => block_number,
 							Err(e) => {
 								log::error!("offchain_worker error: {:?}", e);
@@ -348,8 +341,11 @@ pub mod pallet {
 							},
 						};
 
-						let total_supply = match Self::parse_token_balance(
-							Self::fetch_token_total_supply(token_address.clone(), None),
+						let total_supply = match T::OffchainHttpService::parse_token_balance(
+							T::OffchainHttpService::fetch_token_total_supply(
+								token_address.clone(),
+								None,
+							),
 						) {
 							Ok(total_supply) => total_supply,
 							Err(e) => {
@@ -359,11 +355,13 @@ pub mod pallet {
 							},
 						};
 
-						match Self::parse_token_balance(Self::fetch_token_balance_of(
-							token_address,
-							account_id,
-							None,
-						)) {
+						match T::OffchainHttpService::parse_token_balance(
+							T::OffchainHttpService::fetch_token_balance_of(
+								token_address,
+								account_id,
+								None,
+							),
+						) {
 							Ok(token_balance) => {
 								call = Some(Call::approve_propose {
 									dao_id,
@@ -393,11 +391,13 @@ pub mod pallet {
 						account_id,
 						hash,
 						block_number,
-					} => match Self::parse_token_balance(Self::fetch_token_balance_of(
-						token_address,
-						account_id,
-						Some(block_number),
-					)) {
+					} => match T::OffchainHttpService::parse_token_balance(
+						T::OffchainHttpService::fetch_token_balance_of(
+							token_address,
+							account_id,
+							Some(block_number),
+						),
+					) {
 						Ok(token_balance) => {
 							call = Some(Call::approve_proposal_vote {
 								dao_id,
@@ -831,177 +831,6 @@ pub mod pallet {
 					.collect::<Vec<u8>>()
 			})
 		}
-
-		fn parse_block_number(response: Result<EthRPCResponse, Error<T>>) -> Result<u32, Error<T>> {
-			let result = response?.result;
-			let value = Result::unwrap_or(str::from_utf8(&result), "");
-			if value.is_empty() {
-				return Err(Error::<T>::HttpFetchingError)
-			}
-
-			let value_stripped = value.strip_prefix("0x").unwrap_or(value);
-			let block_number = Result::unwrap_or(u32::from_str_radix(value_stripped, 16), 0_u32);
-
-			Ok(block_number)
-		}
-
-		fn parse_token_balance(
-			response: Result<EthRPCResponse, Error<T>>,
-		) -> Result<u128, Error<T>> {
-			let result = response?.result;
-
-			let value = Result::unwrap_or(str::from_utf8(&result), "");
-			let value_stripped = value.strip_prefix("0x").unwrap_or(value);
-
-			let token_balance = Result::unwrap_or(u128::from_str_radix(value_stripped, 16), 0_u128);
-
-			Ok(token_balance)
-		}
-
-		fn fetch_token_balance_of(
-			token_address: Vec<u8>,
-			account_id: Vec<u8>,
-			block_number: Option<u32>,
-		) -> Result<EthRPCResponse, Error<T>> {
-			let to = str::from_utf8(&token_address[..]).expect("Failed to convert token address");
-
-			let data = [
-				ERC20_TOKEN_BALANCE_OF_SIGNATURE_PREFIX,
-				str::from_utf8(&account_id[..]).map_err(|_| Error::<T>::InvalidInput)?,
-			]
-			.concat();
-
-			let params = json!([
-				{
-					"to": to,
-					"data": data
-				},
-				Self::block_number(block_number)
-			]);
-
-			Self::fetch_from_eth(token_address, None, Some(params))
-		}
-
-		fn fetch_token_total_supply(
-			token_address: Vec<u8>,
-			block_number: Option<u32>,
-		) -> Result<EthRPCResponse, Error<T>> {
-			let to = str::from_utf8(&token_address[..]).expect("Failed to convert token address");
-			let params = json!([
-				{
-					"to": to,
-					"data": ERC20_TOKEN_TOTAL_SUPPLY_SIGNATURE
-				},
-				Self::block_number(block_number)
-			]);
-			Self::fetch_from_eth(token_address, None, Some(params))
-		}
-
-		fn fetch_from_eth(
-			token_address: Vec<u8>,
-			method: Option<Value>,
-			params: Option<Value>,
-		) -> Result<EthRPCResponse, Error<T>> {
-			let key_suffix = &token_address[..];
-			let key_vec = &[
-				key_suffix,
-				&b"::"[..2],
-				&serde_json::to_vec(&method.clone().unwrap_or_else(|| json!(""))).map_err(|e| {
-					log::error!("method parse error: {:?}", e);
-					<Error<T>>::HttpFetchingError
-				})?[..],
-				&serde_json::to_vec(&params.clone().unwrap_or_else(|| json!(""))).map_err(|e| {
-					log::error!("params parse error: {:?}", e);
-					<Error<T>>::HttpFetchingError
-				})?[..],
-			]
-			.concat()[..];
-			let key = [&b"societal-dao::eth::"[..19], key_vec].concat();
-			let s_info = StorageValueRef::persistent(&key[..]);
-
-			if let Ok(Some(eth_rpc_response)) = s_info.get::<EthRPCResponse>() {
-				return Ok(eth_rpc_response)
-			}
-
-			let lock_key = [&b"societal-dao::eth::lock::"[..25], key_vec].concat();
-			let mut lock = StorageLock::<BlockAndTime<Self>>::with_block_and_time_deadline(
-				&lock_key[..],
-				LOCK_BLOCK_EXPIRATION,
-				offchain::Duration::from_millis(LOCK_TIMEOUT_EXPIRATION),
-			);
-
-			if let Ok(_guard) = lock.try_lock() {
-				let json = json!({
-					"jsonrpc": "2.0",
-					"method": method.unwrap_or_else(|| json!("eth_call")),
-					"id": 1,
-					"params": params.unwrap_or_else(|| json!([]))
-				});
-
-				let body = &serde_json::to_vec(&json).expect("Failed to serialize")[..];
-
-				match Self::fetch_n_parse(vec![body]) {
-					Ok(eth_rpc_response) => {
-						s_info.set(&eth_rpc_response);
-
-						return Ok(eth_rpc_response)
-					},
-					Err(err) => return Err(err),
-				}
-			}
-
-			Err(<Error<T>>::HttpFetchingError)
-		}
-
-		fn fetch_n_parse(body: Vec<&[u8]>) -> Result<EthRPCResponse, Error<T>> {
-			let resp_bytes = Self::fetch_from_remote(body).map_err(|e| {
-				log::error!("fetch_from_remote error: {:?}", e);
-				<Error<T>>::HttpFetchingError
-			})?;
-
-			let resp_str =
-				str::from_utf8(&resp_bytes).map_err(|_| <Error<T>>::HttpFetchingError)?;
-
-			let eth_rpc_response: EthRPCResponse =
-				serde_json::from_str(resp_str).map_err(|_| <Error<T>>::HttpFetchingError)?;
-
-			Ok(eth_rpc_response)
-		}
-
-		fn fetch_from_remote(body: Vec<&[u8]>) -> Result<Vec<u8>, Error<T>> {
-			let eth_rpc_url = &EthRpcUrl::<T>::get().to_vec()[..];
-			let request = offchain::http::Request::post(str::from_utf8(eth_rpc_url).unwrap(), body);
-
-			// Keeping the offchain worker execution time reasonable, so limiting the call to be
-			// within 6s.
-			let timeout = sp_io::offchain::timestamp()
-				.add(offchain::Duration::from_millis(FETCH_TIMEOUT_PERIOD));
-
-			let pending = request
-				.deadline(timeout) // Setting the timeout time
-				.send() // Sending the request out by the host
-				.map_err(|_| <Error<T>>::HttpFetchingError)?;
-
-			let response = pending
-				.try_wait(timeout)
-				.map_err(|_| <Error<T>>::HttpFetchingError)?
-				.map_err(|_| <Error<T>>::HttpFetchingError)?;
-
-			if response.code != 200 {
-				log::error!("Unexpected http request status code: {}", response.code);
-				return Err(<Error<T>>::HttpFetchingError)
-			}
-
-			Ok(response.body().collect::<Vec<u8>>())
-		}
-
-		fn block_number(block_number: Option<u32>) -> Value {
-			json!(match block_number {
-				None => "latest".into(),
-				Some(block_number) =>
-					format!("{}{}", "0x", &hex::encode(block_number.to_be_bytes())[2..]),
-			})
-		}
 	}
 }
 
@@ -1120,6 +949,12 @@ impl<T: Config> BlockNumberProvider for Pallet<T> {
 	type BlockNumber = T::BlockNumber;
 	fn current_block_number() -> Self::BlockNumber {
 		<frame_system::Pallet<T>>::block_number()
+	}
+}
+
+impl<T: Config> EthRpcProvider for Pallet<T> {
+	fn get_rpc_url() -> Vec<u8> {
+		EthRpcUrl::<T>::get().to_vec()
 	}
 }
